use CTEs for a single write to PG
