+++ packages/server/graphql/mutations/helpers/createTeamAndLeader.ts
@@ -1,3 +1,5 @@
+import {sql} from 'kysely'
+import TeamMemberId from '../../../../client/shared/gqlIds/TeamMemberId'
 import getRethink from '../../../database/rethinkDriver'
 import MeetingSettingsAction from '../../../database/types/MeetingSettingsAction'
 import MeetingSettingsPoker from '../../../database/types/MeetingSettingsPoker'
@@ -7,8 +9,6 @@ import TimelineEventCreatedTeam from '../../../database/types/TimelineEventCreat
 import {DataLoaderInstance} from '../../../dataloader/RootDataLoader'
 import getKysely from '../../../postgres/getKysely'
 import IUser from '../../../postgres/types/IUser'
-import addTeamIdToTMS from '../../../safeMutations/addTeamIdToTMS'
-import insertNewTeamMember from '../../../safeMutations/insertNewTeamMember'
 
 interface ValidNewTeam {
   id: string
@@ -24,7 +24,7 @@ export default async function createTeamAndLeader(
   dataLoader: DataLoaderInstance
 ) {
   const r = await getRethink()
-  const {id: userId} = user
+  const {id: userId, picture, preferredName, email} = user
   const {id: teamId, orgId} = newTeam
   const organization = await dataLoader.get('organizations').loadNonNull(orgId)
   const {tier, trialStartDate} = organization
@@ -44,14 +44,29 @@ export default async function createTeamAndLeader(
   const pg = getKysely()
   await Promise.all([
     pg
-      .with('Team', (qc) => qc.insertInto('Team').values(verifiedTeam))
+      .with('TeamInsert', (qc) => qc.insertInto('Team').values(verifiedTeam))
+      .with('UserUpdate', (qc) =>
+        qc
+          .updateTable('User')
+          .set({tms: sql`arr_append_uniq("tms", ${teamId})`})
+          .where('id', '=', userId)
+      )
+      .with('TeamMemberUpsert', (qc) =>
+        qc.insertInto('TeamMember').values({
+          id: TeamMemberId.join(teamId, userId),
+          teamId,
+          userId,
+          picture,
+          preferredName,
+          email,
+          isLead: true,
+          openDrawer: 'manageTeam'
+        })
+      )
       .insertInto('TimelineEvent')
       .values(timelineEvent)
       .execute(),
     // add meeting settings
-    r.table('MeetingSettings').insert(meetingSettings).run(),
-    // denormalize common fields to team member
-    insertNewTeamMember(user, teamId, dataLoader),
-    addTeamIdToTMS(userId, teamId)
+    r.table('MeetingSettings').insert(meetingSettings).run()
   ])
 }
